+++ test/Z/IO/FileSystemSpec.hs
@@ -49,6 +49,11 @@ spec = describe "filesystem operations" $ do
                 written <- readExactly size i
                 written @?= content
 
+                fr <- newFilePtr file 0
+                i <- newBufferedInput 4096 fr
+                written <- readExactly size i
+                written @=? content
+
 
             unlink filename
 
@@ -67,6 +72,12 @@ spec = describe "filesystem operations" $ do
                 i <- newBufferedInput 4096 file
                 Just firstLine <- readLine i
                 firstLine  @?= fst (V.break (== V.c2w '\n') content2)
+
+                fr <- newFilePtr file (fromIntegral $ size2 `div` 2)
+                i <- newBufferedInput 4096 fr
+                replicateM_ 512 $ do
+                    Just firstLine <- readLine i
+                    firstLine  @=? fst (V.break (== V.c2w '\n') content2)
 
             unlink filename