+++ ChangeLog.md
@@ -1,5 +1,10 @@
 # Revision history for Z-IO
 
+## 0.1.2.0  -- 2020-09-28
+
+* Add file offset interface back, see `FilePtr` and `FilePtrT`.
+* Remove `checkFileTClosed` from `Z.IO.FileSystem.Threaded`.
+
 ## 0.1.1.2  -- 2020-09-25
 
 * Fix macOS build problem caused by missing zconf.h.